- min energy changed to float

+++ A23 Ex03 DmitryChernyak 307694745 TzahFleischer 305152969/Ex03.GarageLogic/ElectricEngine.cs
@@ -12,7 +12,7 @@ namespace Ex03.GarageLogic
         public float CurrentEnergy { get; set; }
         public float MaxEnergy { get; set; }
         public EnergyType CurrentEnergyType => EnergyType.Electric;
-        public int MinEnergy => 10;
+        public float MinEnergy => 10;
         public float EnergyPercentage => CurrentEnergy / MaxEnergy * 100;
