+++ A23 Ex03 DmitryChernyak 307694745 TzahFleischer 305152969/Ex03.GarageLogic/IEngine.cs
@@ -9,7 +9,7 @@ namespace Ex03.GarageLogic
     {
         float CurrentEnergy { get; set; }
         float MaxEnergy { get; set; }
-        int MinEnergy { get; }
+        float MinEnergy { get; }
         EnergyType CurrentEnergyType { get; }
         float EnergyPercentage { get; }
